BL-8933 Installer link component
* for CS version link
