+++ src/components/ContentfulPage.tsx
@@ -7,7 +7,8 @@ import { documentToReactComponents } from "@contentful/rich-text-react-renderer"
 import { useContentful } from "../connection/UseContentful";
 import { useDocumentTitle } from "./Routes";
 import Markdown from "markdown-to-jsx";
-import { WindowsInstallerDownloads } from "./WindowsInstallerDownloads";
+import { WindowsInstallerDownload } from "./WindowsInstallerDownload";
+import { WindowsInstallerLink } from "./WindowsInstallerLink";
 import { BloomReaderVersionNumber } from "./BloomReaderVersionNumber";
 import { CreationThemeProvider } from "../theme";
 import Link from "@material-ui/core/Link";
@@ -58,8 +59,11 @@ export const ContentfulPage: React.FunctionComponent<{ urlKey: string }> = (
                               a: {
                                   component: Link,
                               },
+                              WindowsInstallerLink: {
+                                  component: WindowsInstallerLink,
+                              },
                               WindowsInstallerDownload: {
-                                  component: WindowsInstallerDownloads,
+                                  component: WindowsInstallerDownload,
                               },
                               BloomReaderVersionNumber: {
                                   component: BloomReaderVersionNumber,
